+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [ 1.1.0 ] - 2022-04-06
+### Changed 
+- Enhanced IEnumConfigurationResourceResolver to load by name.
+- Changed EnumConfigurationValue implementation to support newIterator.
+
 ## [ 1.0.0 ] - 2022-02-18
 ### Added
 - Introduced IEnumConfigurationStore and sample implementation PropertiesEnumConfigurationStore.